requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
flake8==7.3.0
kombu==5.6.2
mccabe==0.7.0
numpy==2.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.1.0
prompt_toolkit==3.0.52
pycodestyle==2.14.0
pyflakes==3.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.1.0
reportlab==4.4.7
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
